fix(hol-guard): address package workbench review feedback
Preserve camelCase advisory IDs, clear stale audit snapshots, wire audit
through a ref callback, split firewall controls into focused modules, and
fix workbench sort state and table semantics.
