+++ src/codex_plugin_scanner/guard/daemon/static/assets/guard-dashboard.js
@@ -25291,14 +25291,14 @@ export {
   exportSettings as a7,
   importSettings as a8,
   resetSettings as a9,
-  runPackageAudit as aA,
-  runPackageSync as aB,
-  startPackageFirewallConnect as aC,
-  openPackageFirewallShell as aD,
-  HiMiniBugAnt as aE,
-  HiMiniClock as aF,
-  IconActionButton as aG,
-  HiMiniBeaker as aH,
+  HiMiniBeaker as aA,
+  HiMiniBugAnt as aB,
+  fetchPackageFirewallStatus as aC,
+  runPackageFirewallAction as aD,
+  runPackageAudit as aE,
+  runPackageSync as aF,
+  startPackageFirewallConnect as aG,
+  openPackageFirewallShell as aH,
   HiMiniArrowDown as aI,
   HiMiniArrowUp as aJ,
   fetchSupplyChainBundle as aK,
@@ -25333,8 +25333,8 @@ export {
   Surface as av,
   HiMiniArrowTopRightOnSquare as aw,
   HiMiniCheckBadge as ax,
-  fetchPackageFirewallStatus as ay,
-  runPackageFirewallAction as az,
+  HiMiniClock as ay,
+  IconActionButton as az,
   HiMiniExclamationTriangle as b,
   HiMiniArrowRight as c,
   HiMiniChevronUp as d,
